Added Bubbles pickup prefab. Changed controls.
2:00 PM -6:00 PM : Did some reading up on colliders, events, rigid body,
unity physics, and that sort of thing. Found out that the passthrough
was because the existing colliders were being ignored, and that some of
the problems were because I was using transform to move the player
object around instead of Forces. Changed the code to apply forces
instead of manually setting the transform. Added scalars to adjust
movement speeds when boosting or breaking. Added inputs to use Shift and
Ctrl to change those scalars (They correspond to the triggers on an Xbox
controller, but the xbox controller inputs have not been configured or
tested with the input manager.) Finally got the rigid body working
properly, and event collider for the bubbles pickup working properly.
Game Over is still not enforced, but it should be trivial to add. Now we
should start adding enemies and any other pickups similar to the
Bubbles.