Cursor is now hidden and locked to center screen.
Removed a deleteme line. Added a new GameController script. Added a
state machine to the GameController script. Added a Game Controller
Object to the scene and attached the script to it. Currently, it only
checks to see if the gamestate is 'playing' and if so, it locks the
cursor. If the state is not set in the inspector by Start(), the state
defaults to Playing. Read more about screenlocking here:
http://answers.unity3d.com/questions/62440/how-to-stop-mouse-going-out-of-the-screen-.html
http://docs.unity3d.com/Documentation/ScriptReference/Screen-lockCursor.html
Added a note to the readme.md to remind the user how to stop gameplay
without being able to click the pause button.