Placeholder Enemy Deals Damage
Added some textures for the Slabs, Boogers, and Stalks. Added a
placeholder enemy that can damage the player if the player is looking at
it. Put the collision detection cone onto the player instead of the
enemies. Added scripts to hander enemy behaviors and player behaviors.
Slightly tweaked player movement.