Enemies Now Flee Player
I tried to move a bunch of utility methods into a separate classfile,
LibRevel.cs. I apparently broke that though, and now enemies flee away
from the player at superspeed.

The TriggerEnter wasn't detecting the player ramming against the enemy.
Instead, I'll work around it for now by giving a distance check for a
very small amount of distance.

Also, it's likely that the raycasting is being projected from the center
of gameobjects instead of the bounds. In which ase, we should probably
factor in the radius from the center of enemies to its surface as an
additional factor of the length it needs to travel to reach a target.