Created new models DEYE, Booger, and Stalk.
The Curved stalk file has an anim in which it curves and inflates as if
it's adigtated. The straight stalk has no such animation.

The DEYE is an enemy to chase the player if they stay in one area too
long. It will pursue the player slowly but relentlessly, deals a lot of
damage on look or on touch, and can not be slain. If the player leaves
to a new terrain tile, the DEYE will disappear.

Added model for Booger.  Added stretch animation for Booger.

Updated Readme.md.

Default testing scene now uses the new Stalk model for testing the Stalk
enemy, but it is not a prefab.